# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SourceTrace API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/knowledge-bases/{knowledge_base_id}/documents": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Upload Document",
        "operationId": "upload_document_api_v1_knowledge_bases__knowledge_base_id__documents_post",
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Knowledge Base Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_knowledge_bases__knowledge_base_id__documents_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              }
            }
          },
          "200": {
            "description": "Exact duplicate resolved to the existing document version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentUploadResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "413": {
            "description": "Content Too Large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unsupported Media Type"
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      },
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "List Document Versions",
        "operationId": "list_document_versions_api_v1_knowledge_bases__knowledge_base_id__documents_get",
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Knowledge Base Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentVersionListResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/knowledge-bases/{knowledge_base_id}/documents/{version_id}/retry": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Retry Document Ingestion",
        "operationId": "retry_document_ingestion_api_v1_knowledge_bases__knowledge_base_id__documents__version_id__retry_post",
        "parameters": [
          {
            "name": "knowledge_base_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Knowledge Base Id"
            }
          },
          {
            "name": "version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Version Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestionRetryResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_document_api_v1_knowledge_bases__knowledge_base_id__documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_knowledge_bases__knowledge_base_id__documents_post"
      },
      "DocumentUploadResponse": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Version Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version_number": {
            "type": "integer",
            "title": "Version Number"
          },
          "checksum_sha256": {
            "type": "string",
            "title": "Checksum Sha256"
          },
          "file_size_bytes": {
            "type": "integer",
            "title": "File Size Bytes"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "status": {
            "type": "string",
            "maxLength": 24,
            "minLength": 1,
            "title": "Status"
          },
          "stage": {
            "type": "string",
            "maxLength": 24,
            "minLength": 1,
            "title": "Stage"
          },
          "attempt_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Attempt Count"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable"
          },
          "failure_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Code"
          },
          "failure_message": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "deduplicated": {
            "type": "boolean",
            "title": "Deduplicated"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "version_id",
          "name",
          "version_number",
          "checksum_sha256",
          "file_size_bytes",
          "page_count",
          "status",
          "stage",
          "attempt_count",
          "retryable",
          "created_at",
          "deduplicated",
          "request_id"
        ],
        "title": "DocumentUploadResponse"
      },
      "DocumentVersionItem": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Version Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version_number": {
            "type": "integer",
            "title": "Version Number"
          },
          "checksum_sha256": {
            "type": "string",
            "title": "Checksum Sha256"
          },
          "file_size_bytes": {
            "type": "integer",
            "title": "File Size Bytes"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "status": {
            "type": "string",
            "maxLength": 24,
            "minLength": 1,
            "title": "Status"
          },
          "stage": {
            "type": "string",
            "maxLength": 24,
            "minLength": 1,
            "title": "Stage"
          },
          "attempt_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Attempt Count"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable"
          },
          "failure_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Code"
          },
          "failure_message": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "version_id",
          "name",
          "version_number",
          "checksum_sha256",
          "file_size_bytes",
          "page_count",
          "status",
          "stage",
          "attempt_count",
          "retryable",
          "created_at"
        ],
        "title": "DocumentVersionItem"
      },
      "DocumentVersionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentVersionItem"
            },
            "type": "array",
            "title": "Items"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "DocumentVersionListResponse"
      },
      "ErrorResponse": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "instance": {
            "type": "string",
            "title": "Instance"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance",
          "code",
          "request_id"
        ],
        "title": "ErrorResponse"
      },
      "FieldError": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "code": {
            "type": "string",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "field",
          "message",
          "code"
        ],
        "title": "FieldError"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestionRetryResponse": {
        "properties": {
          "version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Version Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "attempt_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Attempt Count"
          },
          "retryable": {
            "type": "boolean",
            "title": "Retryable"
          },
          "failure_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Failure Code"
          }
        },
        "type": "object",
        "required": [
          "version_id",
          "status",
          "stage",
          "attempt_count",
          "retryable"
        ],
        "title": "IngestionRetryResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}